backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pre-commit==3.6.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.21
aiofiles==23.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
slowapi==0.1.9
email-validator==2.3.0
asyncpg
pyotp
qrcode
redis
tenacity==8.2.3  # For retry logic with exponential backoff

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# DICOM Processing
pydicom==3.0.1
dicomweb-client==0.60.1
pillow==12.1.0

# Storage
boto3==1.42.25
botocore==1.42.25

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
httpx==0.28.1

# Code Quality
ruff==0.14.11
black==25.12.0
mypy==1.19.1
reportlab==4.2.5
sse-starlette==2.2.1

# Error Monitoring
sentry-sdk[fastapi]==2.25.1

# Monitoring/Metrics
prometheus-fastapi-instrumentator==7.0.0
apscheduler==3.10.4
prometheus-fastapi-instrumentator==7.0.0
